backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.50.0
playwright==1.47.0
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic==2.9.2
httpx>=0.27.2
python-multipart==0.0.12
slowapi==0.1.9
supabase>=2.10.0
